split schema form
vibe coding
